Removing silly comments and prints
