+++ .gitignore
@@ -5,3 +5,4 @@ db.py
 .idea
 .DS_Store
 /.vscode
+.pytest_cache
