+++ app/config.py
@@ -3,6 +3,7 @@ from dotenv import load_dotenv
 
 load_dotenv()
 class Config(object):
+    
     SQLALCHEMY_DATABASE_URI = os.getenv('DB_CON_STR')
     SQLALCHEMY_TRACK_MODIFICATIONS = False
     SECRET_KEY = 'Secure'
@@ -11,9 +12,6 @@ class Config(object):
     DEBUG = True
     SQLALCHEMY_TRACK_MODIFICATIONS = False
     
-
-
-
 class TestingConfig(Config):
     
     SQLALCHEMY_DATABASE_URI = os.getenv('TEST_DB_CON_STR')
